Move fontinfo data into own module
